Move PBL calc to new header file
